Optimized WorldManager.EnumerateIntersectionObjects
According to the Visual Studio profiler (on release mode), the game was
spending about 26% of execution time in
Microsoft.XMA.Framework.BoundingBox.Intersect(BoundingFrustum) playing a
small map with about 5 dwarves and quite a bit of wildlife.

This method is only used by the renderer to enumerate visible bodies in a
recursive frustum-AABB intersection test.
Besides the method itself which uses a seemingly unbounded loop, there
might be some calling overhead due to the method lying in a external assembly.

I replaced it with a more optimized version called OctTreeNode.FastIntersects
and added a version of OctTreeNode.EnumerateItems with a Filter paramter
to more aggresively cull bodies.
This redueces the time spent by EnumerateIntersectionObjects to 8%,
and seems to improve the FPS and reduce jitter.

As described in http://old.cescg.org/CESCG-2002/DSykoraJJelinek/
there is room for further improvement, such as using plane masking
(where if we know that an octree lies one side of a plane, then all
its children also lie that side of the plane so we don't need to test again),
or some sort of frame-to-frame coherency where the results of the last
frame are used to speed up the body culling of the current frame.